fix delete event

+++ api/src/routes/controllers/deleteControllers.js
@@ -38,14 +38,17 @@ const deleteGroups = async (req, res) => {
 const deleteEvent = async (req, res) => {
   const { id } = req.params;
   try {
-    await Event.destroy({
-      where: {
-        id,
-      },
-    });
-    res.json({ message: "Event deleted" });
+    if (!id) { res.status(400).json({error: "missing id"})
+    } else { 
+      await Event.destroy({
+        where: {
+          id,
+        },
+      });
+      res.json({ message: "Event deleted" });
+    }
   } catch (error) {
-    res.json(error);
+    res.status(400).json({error_DB: error.message});
   }
 };
